refactor: fold report_sync_result into absurd_sync_queues
AbsurdReportCommand existed to share queue-SyncResult reporting between
absurd_sync_queues and absurd_worker. The worker no longer reports one, leaving a base
class with a single subclass and nothing to share. absurd_sync_crons is not a candidate
for it: cron sync returns two counts and prunes, where SyncResult carries lists of queue
names and storage-mode warnings.

The method moves to the command that uses it, and its parameters lose the defaults no
caller ever took.
